+++ src/protocols/mod.rs
@@ -4,5 +4,7 @@ pub mod nec;
 pub mod rc5;
 #[cfg(feature = "rc6")]
 pub mod rc6;
+#[cfg(feature = "sbp")]
+pub mod sbp;
 
 pub mod logging;