+++ src/remotes/mod.rs
@@ -5,6 +5,7 @@ pub use remotecontrol::*;
 pub mod nec;
 pub mod rc5;
 pub mod rc6;
+pub mod sbp;
 
 
 #[cfg(feature = "std")]