+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## [Unreleased] - 2019-10-XX
+## [Unreleased] - 2019-11-XX
  - Added (optional) embedded-hal support
  - Added support for NEC commands with 16 bit addresses
+ - Added support for Samsung BluRay Player protocol
 
 ## [0.4.1] - 2019-09-29
  - Added support for Rc5 transmit